Added CSE code in anormalize-cse, exptable.* Also rearranged structures in nilopts* so that I have more control over the optimizations from the command line. More specifically, the flags NilOpts.do_reduce NilOpts.do_flatten and NilOpts.do_cse are there.
Also there is code to keep statistics on how many optimizations are
done. I would use the code in Stats, but they are reset every time the
compiler is run, and I don't want to make them local to the functions
that traverse the modules.

Finally reduce has code to turn project_sum and then select to just
prject_sum_record.

-SCW



Adding hooks for optimizer. --SCW