Improved the hash function for a slight speedup.
Eliminated the hand-written partition code in insert_edges in favor of
List.partition.  I did this to clean up the code but it gave a
surprising speedup.  The ratio of times spent in insert_edges while
compiling toil.sml after/before the change is 192.37s/210.24s = 0.92.
I guess this speedup is due to the fact that the hand-written code
used an accumulator that was a pair of lists while List.partition
(presumably) uses a pair of accumulators.