landing my pattern compiler changes for cleanup week. Visible outside this module:
case x of
   ref => 1

.. is now rejected with a proper error message.

Now raise "Bind" instead of "Match" for
val 0 = 1

Now print warnings for nonexhaustive matches and bindings (not sure this is perfect).

Now use error_region more often so that error messages are less obscure; changed
the text of some really brief ones to be more suggestive of the actual problem.