Moved imports from source files to group files.
TiltPrim unit is now compiled to assembler.  This means that we can do
away with Firstlude and define bool in TiltPrim.

TiltPrim.littleEndian has gone away; it is replaced by a group file variable.
(We did not want this platform-specific value inlined in TiltPrim's
interface.)

Equality primitives for numeric types were added to TiltPrim.  The unit that
defines vector equality no longer has to invoke the equality compiler.