Moved imports from source files to group files.
Passing a version of con_eqfun to the equality compiler when we are
defining bool.  The IlUtil version now looks bool up by unit name and
type name, so it does not function while elaborating the unit that
defines bool.

Wrapping some functions with UtilError.dontShow to suppress spurious
error messages.  (Perhaps we should move the printing from UtilError
to ExnHandler to eliminate the need for this cruft.)