Moved imports from source files to group files.
The labels defined by "extern" declarations are now separate from
other sdecs.  Externs are part of the ad hoc component of an
interface.  "New" externs are not passed to the phase splitter; this
leads to a restriction whereby a unit may not refer to any externs it
defines.  (No such restriction exists for signatures because "new"
signature definitions are expanded prior to phase splitting.)
Introduced CONTEXT_EXTERN and PHRASE_CLASS_EXT.  These carry two
labels, one that may be replaced by a dummy label by shadowing and
another one that does not change and is used to refer to the extern.

The memoized signatures in elaboration contexts are now attached to
context entries rather than stored in a separate table.
PHRASE_CLASS_MOD now carries a unit -> signat.

Replaced partial contexts with parameterized interfaces.  There are
three kinds of parameters.  A parameter can refer to a unit's main
module, to one of its signatures, or to one of its externs.