Improved the project description summaries used for IR to use source interface equivalence when possible.
The gensyms were being reset too early to make compiled interface
files consistent.  They are now being reset just before the elaborator
is called (after, for example, all pinterfaces have been read in
through the file cache).

The temporary files used for linking executables are relative to
the current directory rather than the first project description
file on the command line (which could be a library that should
not be modified).