My last commit claimed to fix an old bug in the sparc generational collectors.  That claim was false: The symptoms went away but the bug did not.  This commit makes up for it.
Changed the C call interface and the C code supporting the basis.  The
sparc GC does not know about C roots.  (That is the old bug.) The rule
now is that sparc C calls may perform at most one allocation and must
consider any other ML pointers on hand to be invalidated by
allocation.  Changed the allocator used by sparc C calls to detect any
second allocation and abort.

Changed the sparc runtime's paranoid code.  There was a concurrent GC
invariant that does not hold in the non-concurrent collector.  The
paranoid code now only checks for that invariant in the concurrent
case.